// types.ts
export interface Map {
    category: string;
    coverImg: string;
    date: string;
    downloadBlackAndWhiteUrl: string;
    downloadColorUrl: string;
    excerpt: string;
    lore: string;
    note: string;
    slug: string;
    subCategory?: string;
    thumbnailUrl: string;
    title: string;
}